2833: Refactor components and fix tests
